third_party/blink/web_tests/http/tests/websocket/workers/no-subprotocol-expected.txt

WebSocket's protocol attribute should be an empty string if the server did not provide Sec-WebSocket-Protocol header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS PASS: ws.protocol is equal to ""
INFO: Connected
PASS PASS: ws.protocol is equal to ""
INFO: Closed
PASS PASS: ws.protocol is equal to ""
PASS PASS: closeEvent.wasClean is true
INFO: Exited onclose handler
PASS PASS: ws.protocol is equal to ""
DONE
PASS successfullyParsed is true

TEST COMPLETE
